# frontend/11心理健康聊天机器人 API.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "11心理健康聊天机器人 API",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/user/emotion-profile": {
      "get": {
        "summary": "Get User Emotion Profile",
        "deprecated": false,
        "description": "获取当前用户的情绪画像",
        "operationId": "get_user_emotion_profile_api_user_emotion_profile_get",
        "tags": [
          "User Profile"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{access_token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get User Emotion Profile Api User Emotion Profile Get",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/emotion-stats": {
      "get": {
        "summary": "Get User Emotion Stats",
        "deprecated": false,
        "description": "获取用户情绪统计信息",
        "operationId": "get_user_emotion_stats_api_user_emotion_stats_get",
        "tags": [
          "User Profile"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "Bearer {{access_token}}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get User Emotion Stats Api User Emotion Stats Get",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "bearer": []
    }
  ]
}